move base class and primary_key type to sqlalchemy.py file
