+++ app/modules/amap/models_amap.py
@@ -6,9 +6,8 @@ from sqlalchemy import Date, Enum, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.models_core import CoreUser
-from app.database import Base
 from app.modules.amap.types_amap import AmapSlotType, DeliveryStatusType
-from app.types.sqlalchemy import TZDateTime
+from app.types.sqlalchemy import Base, TZDateTime
 
 
 class AmapOrderContent(Base):
