+++ app/modules/calendar/models_calendar.py
@@ -4,9 +4,8 @@ from sqlalchemy import Boolean, Enum, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.models_core import CoreUser
-from app.database import Base
 from app.modules.calendar.types_calendar import CalendarEventType
-from app.types.sqlalchemy import TZDateTime
+from app.types.sqlalchemy import Base, TZDateTime
 
 
 class Event(Base):
