+++ app/modules/loan/models_loan.py
@@ -4,7 +4,7 @@ from sqlalchemy import TEXT, Boolean, Date, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core import models_core
-from app.database import Base
+from app.types.sqlalchemy import Base
 
 
 class Loaner(Base):
