+++ app/modules/phonebook/models_phonebook.py
@@ -1,8 +1,8 @@
 from sqlalchemy import Enum, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.database import Base
 from app.modules.phonebook.types_phonebook import Kinds
+from app.types.sqlalchemy import Base
 
 
 class Membership(Base):
